requirements.txt
streamlit==1.54.0
pillow==12.1.0
requests==2.32.5
geopy==2.4.1
serpapi==0.1.5
google-genai==1.62.0
tavily==1.1.0
python-dotenv==1.2.1
